pyproject.toml
[project]
name = "clawworld"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    "httpx",
    "click",
    "jinja2",
    "python-dotenv",
]

[project.scripts]
clawworld = "cli.clawworld_cli:main"

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["cli*", "gateway*"]
